[BE] redactor: #61 application.yml 소셜(kakao, naver, google) 정의 수정 각 소셜 properties 수정
