--- backend/src/main/java/com/ballx/config/properties/GoogleSocialProperties.java
@@ -1,27 +1,0 @@
-package com.ballx.config.properties;
-
-import java.util.Set;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "spring.security.oauth2.client.registration.google")
-public record GoogleSocialProperties(
-	String authorizationGrantType,
-	String clientAuthenticationMethod,
-	String clientId,
-	String clientName,
-	String clientSecret,
-	String redirectUri,
-	Set<String> scope
-) {
-	public record Provider(
-		String authorizationUri,
-		String issuerUri,
-		String jwkSetUri,
-		String tokenUri,
-		String UserInfoAuthenticationMethod,
-		String userInfoUri,
-		String userNameAttribute
-	) {
-	}
-}
