+++ backend/src/main/java/com/ballx/config/properties/social/KakaoSocialProperties.java
@@ -1,0 +1,13 @@
+package com.ballx.config.properties.social;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "oauth.kakao")
+public record KakaoSocialProperties(
+	String clientId,
+	String clientSecret,
+	String redirectUri,
+	String tokenUrl,
+	String userInfoUrl
+) {
+}
