+++ backend/src/main/java/com/ballx/config/properties/social/NaverSocialProperties.java
@@ -1,0 +1,13 @@
+package com.ballx.config.properties.social;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "oauth.naver")
+public record NaverSocialProperties(
+	String clientId,
+	String clientSecret,
+	String redirectUri,
+	String tokenUrl,
+	String userInfoUrl
+) {
+}
