+++ backend/src/main/resources/application.yml
@@ -61,24 +61,23 @@ sms:
 
 oauth:
   kakao:
+    client-id: ${KAKAO_CLIENT}
+    client-secret: ${KAKAO_SECRET}
     redirect: ${KAKAO_REDIRECT_URI}
-    client: ${KAKAO_CLIENT}
-    secret: ${KAKAO_SECRET}
-    token-uri: https://kauth.kakao.com/oauth/token
-    user-info-uri: https://kapi.kakao.com/v2/user/me
+    token-url: ${}https://kauth.kakao.com/oauth/token
+    user-info-url: https://kapi.kakao.com/v2/user/me
   naver:
-    redirect: ${NAVER_REDIRECT_URI}
-    client: ${NAVER_CLIENT}
-    secret: ${NAVER_SECRET}
-    token-uri: https://nid.naver.com/oauth2.0/token
-    user-info-uri: https://openapi.naver.com/v1/nid/me
+    client-id: ${NAVER_CLIENT}
+    client-secret: ${NAVER_SECRET}
+    redirect-uri: ${NAVER_REDIRECT_URI}
+    token-url: https://nid.naver.com/oauth2.0/token
+    user-info-url: https://openapi.naver.com/v1/nid/me
   google:
-    redirect: ${GOOGLE_REDIRECT_URI}
-    client: ${GOOGLE_CLIENT}
-    secret: ${GOOGLE_SECRET}
-    token-uri: https://oauth2.googleapis.com/token
-    user-info-uri: https://www.googleapis.com/oauth2/v3/userinfo
-    # authorization-uri: https://accounts.google.com/o/oauth2/v2/auth
+    client-id: ${GOOGLE_CLIENT}
+    client-secret: ${GOOGLE_SECRET}
+    redirect-uri: ${GOOGLE_REDIRECT_URI}
+    token-url: https://oauth2.googleapis.com/token
+    user-info-url: https://www.googleapis.com/oauth2/v3/userinfo
 
 rest-client:
   connect-timeout-seconds: 5
